ai_module/requirements.txt
ultralytics
opencv-python
python-socketio[client]
requests
python-dotenv
numpy
shapely
